arm/smmuv3: Track count instead of individual bindings
Again the motivation is to avoid dynamic memory allocations.
The downside is that unbinding/deleting of Dmar_space gets less
efficient.

In addition the counter might go out of sync if a binding is overbound.
That is only a performance issue, because it might result in us doing
TLB flushes for a Dmar_space on an IOMMU it actually is no longer bound
to, but we don't know about that.

This also fixes a bug in deconfigure_and_invalidate_tlb(), where we
missed to execute a TLB flush, because the is_bound() check in the
function came before the invocation of del_binding() in unbind().

See ticket #CD-1351.

Change-Id: I74d8fe4022f95a9c0ac02dd51c16976b4049a2e8