Clang older 9 not supported anymore
In theory, Clang>=5 is able to build for IA32/AMD64 with -std=c++17, but
the detection mechanism for CONFIG_CC_HAS_ASM_FLAG_OUTPUTS doesn't work
well because Kconfig uses the host gcc rather than CONFIG_CC to execute
the detection routine in the Kconfig file.

For ARM, at least Clang-9 is required to provide the required "asm goto"
feature.

Change-Id: Icae6fd52acc686930caeafd4b67ba031d388106e