riscv: ld.lld: KEEP logging string identifier strings
The .rodata.log.str section contain strings required for JDB logging. If
this section exists, it must be added to the final binary and the linker
is not allowed to strip it.

Note that code in section `.debug.jdb.log_table` has references to these
strings, hence the linker shall not strip it. For some reason, the LLVM
linker seems to have a different opinion, hence the KEEP statements. For
GNU ld these statements are not required but they don't cause problems.

Addendum to I70988e25fbbe83561e7b30b6f40fc36bd0578053.

Change-Id: I0bf912ba1ec0c7d39fb5c0b7f1301838ee856ae0