show preprocess command line only on V=2
Do not show the detailed preprocess command any longer on "make V=1"
because these commands clutter the screen even if only verbose output
of compiling tools is of interest. Use "make V=2" to show the detailed
preprocess command line.

Also reorganize the preprocess receipt to not inject more newlines in
the output as necessary.

Change-Id: I4ae93b32bf5bc72b6b7bedb863b1b4ae65984432