arm: fix vtimer injection if guest is at WFI
Usually the VMM will trap WFI instructions to go to sleep. It will
return to the guest if either an external event happens (IPC, doorbell
irq, ...) or the virtual generic timer expired.

If the vtimer is pending when going back to the guest, we must inject
it immediately. Interrupts are disabled until we leave the kernel. If
the guest is at a WFI instruction we won't get the vtimer interrupt
because the WFI instruction trap has a higher priority than interrupts!

Change-Id: I656bf70fbd3e31fc83c4ec433e0ea90d12bb7d33