assertion: Avoid code duplication
The mux console is always built in, no matter if this is a JDB build or
not. So the end_exclusive call is always available and there is no harm
on calling this.

Also the system_abort function will default to a terminate(EXIT_FAILURE)
call on non-JDB builds. So the functions are essentially equivalent.
Remove the !debug version to avoid code duplication.

Change-Id: I44388af18ee6ed417ff48cdb4692a067476bb91d