Fix out-of-bounds send item access in sys_register_delete_irq()
The syscall handler Thread_object::sys_register_delete_irq() calls
next() on L4_snd_item_iter without calling more() beforehand. This is
problematic because next() itself does not contain a range check but
relies on the user calling more() beforehand.

Fortunately, the memory area that Fiasco can be tricked into accessing
due to this missing check is limited to the two machine words that are
located in the UTCB after the message registers. Since user mode can
read and write everything in the UTCB anyway, and the message registers
are not subject to any special pre/post check, the out-of-bounds access
cannot cause any harm.

This change makes sure that the kernel does not use UTCB values outside
of the UTCB message registers as send items.

Change-Id: I5386981aa891824e24c65aaec673279028255356