Unify page & superpage arithmetics (no change in functionality)
Improve the readability of the code by replacing the custom
computations of rounded/truncated addresses/sizes, offsets, etc.,
with respect to pages and superpages by the usage of unified methods.

This should not only make the code easier to understand, but should also
prevent from accidentally mixing up some of the more creative
computations for something else.

Change-Id: I82d1eb8e62ec5f3ae64e5fe551cf61cd499f73fd