x86: Avoid crashing on stale TLB entries during application CPU startup
Previously, when CPU local mapping was enabled, the startup code of
application CPUs was reusing the local mapping of the bootstrap CPU
until the proper CPU local mapping was established.

This means that there were, in sequence, two distinct global mappings of
the Kentry_cpu_page installed on the same CPU. Since the "inherited"
mapping was implicitly used by the CPU to access the GDT during the
startup process, it could end up in the TLB and would not be removed
from it during the installation of the "proper" mapping of
Kentry_cpu_page due to the global flag.

This can lead to the incorrect access into the Kentry_cpu_page of the
bootstrap CPU on the application CPU in Kmem::init_cpu() even after the
"proper" mapping was installed. This can consequently lead to the
incorrect installation of the CPU local mapping of the application CPU
on the bootstrap CPU in the next call of Mem_space::switch_page_table().

While the issue has not been observed on bare metal, it is easy to
trigger in QEMU/KVM.

This change runs the startup code of the application CPU with the
non-local kernel mapping (kdir) that does not contain conflicting
global translations with the CPU local mapping, thus eliminating
the possibility of a stale TLB entry.

Since the startup code needs to access the GDT which is originally
allocated in the Kentry_cpu_page in case of CPU local mapping, we
provide the startup code with a static copy of the GDT that is
accessible via the non-local kernel mapping (kdir).

This change fixes #CD-159.

Change-Id: I1e558dc362f39a3ab1992a264d5b729d73f9b6e8