Cpu_call::cpu_call_many(): return early on empty CPU set
If an empty CPU set was passed to Cpu_call::cpu_call_many(), leave the
function early. There was code to achieve this before but it didn't work
because `cs.used()` cannot be 0 after `cs.next()` was successfully
called for at least one time.

Also remove Cpu_call::used() as this function is now superfluous.

Change-Id: I4efe6063b949039a8c4e6154e48957cf2726ab20