task: remove unused code for allocating KU memory >= L4_SUPERPAGESIZE
The caller ensures that the to-be-allocated size of KU memory is never
larger than L4_SUPERPAGESIZE. In fact, the limit forced by the buddy
allocator is 2^17, not 2^20. Hence, Task::alloc_ku_mem_chunk() does not
need to consider this case.

Adapt the limit in the caller.

Also rename `page_size` to `page_order` in alloc_ku_mem_chunk() and in
free_ku_mem_chunk().

Change-Id: I196c557b64f194e31b85ea269f080c69f35b1892