Improve the simple pointer-bumping allocator
Allow the instances of the simple pointer-bumping allocator to either
explictly avoid any locking (as it is currently used within Kmem) or
use a specific locking policy (e.g. locking using a spinlock).

Clearly document (both in the type name and in the comments) that the
allocator instances within Kmem do not use any locking. Any race
conditions are avoided either thanks to a CPU-local usage of the
instances or thanks to the fact that the CPUs are initialized
sequentially.

Improve the API and the usage of the allocator slightly.

Furthermore, Kmem::tss_mem_vm needs to be declared as Static_object<>
to avoid the situation when the default static constructor of
Kmem::tss_mem_vm is accidentally executed after the execution of
Kmem::init_mmu() which configures the proper state of Kmem::tss_mem_vm
(specifically in Kmem::setup_global_cpu_structures()).

This situation when the unspecified order of construction leads to an
invalid Kmem::tss_mem_vm (and further to a crash of the CPU
initialization in Kmem::init_cpu()) was previously avoided only by the
linking order, i.e. basically by chance.

Change-Id: Ie1945016104e4b74fe38f0999f932d4606685ea0