clean up a few panic() messages
Messages printed by panic() should not contain a trailing '\n' because
this is the last message the system prints anyway. The majority of the
messages do this right already.

Also don't mention "halted", "FATAL" or "FIASCO" and let the messages
start with a capital letter.

This change touches only a few messages. Other messages should be fixed
when the corresponding code is touched.

Change-Id: I4684fb69775f42a2d0ddfd8dadfe84f135da495d