utest-fw: Symlink test configs instead of copying
When creating the unit test supplemental files in the builddir symlink
unchanged files instead of copying. This removes the job of the user to
re-generate the builddir on every change to a test configuration.

While at it also clean up the comment for supplemental file parsing that
got outdated when the mechanism got more flexible.

Change-Id: I057d0605bb026cccdbbaac90016923767e0aa897