minilibc: restore standard clamping behavior
To ensure that on 64-bit systems the number returned when exceeding the
value range of a signed long is as reasonably expected we must check
against the correct unsigned values.

While at it also restore the clamping behavior of the function.

Note that this fix has no real correctness or security implications
since the function is only used by printf, all our format strings are
fixed / constant and never even come close to these boundary values.

Change-Id: Ibe5827203cfc50eeaf530fb334c073efa5f1eaa6