riscv: drivers: add inline asm mmio access functions
When running virtualized, the instruction decoding only supports a
subset of instructions forcing us to explicitly use the instructions
we need to.

Note that reading/writing 64-bit on ARM32 is not done using a single
instruction and hence is not atomic.

Addendum to I1c20a628ab96639f1829be57bd844d51046963ea.

Change-Id: I95718d097b9d74dc8bff10c47c2af8827aacc1a9