Allocate permanent kernel structures directly from memory descriptors
The purpose of this change is to make the allocation of the permanent
kernel data structures more generic and less ad hoc (in preparation for
allocating diverse structures in this way in the future).

The permanent allocation is done directly from the memory descriptors
which need to be fixed up temporarily for the purpose of the kernel
memory initialization.

The memory descriptors cannot be altered directly since the original
memory ranges they store are used by sigma0 to identify memory reserved
by the kernel. This is the motivation for the fixup mechanism.

While at it, some relevant argument and variable types have been changed
from unsigned long to Address in order to better express their semantic
meaning. Since Address is currently defined as unsigned long, this does
not change any functionality.

Change-Id: Ie32b030912bcef4c078e816f83f38a3729c03caf