replace assert_opt() for parameters with __attribute__((nonnull()))
The intended use for `assert_opt(arg)` is to tell the compiler that
'!(arg)' cannot be true. And in case this would happen anyway, the
`__assert_fail()` function would be called in case NDEBUG is not
defined.

If an 'assert_opt(arg)' statement is used to show that the function
parameter 'arg' cannot be nullptr, use __attribute__((nonnull())) for
the corresponding parameter as the mechanism that is intended for such
code. For pedantic reasons, still use `assert(arg)` in such cases (makes
only a difference if NDEBUG is not defined).

With NDEBUG enabled, The generated code using this change differs
slightly from the generated code without this change but ultimately it's
only a little re-ordering, likely due to different assumptions about
branch prediction but also a few saved check for nullptr. GCC 13 and
Clang 17 were used to compare the generated code for ARM64 and AMD64.

This change is one part to remove the usage of the `assert_opt()` macro
because that macro does not play well together with code coverage.

Change-Id: I73c1daacfef3f173a003d380b27e27d6165607dc