clarify return of xcpu_state_change and make use of this information
The function `Context::xcpu_state_change()` returns only `true` if the
passed `add` state change bits have any bit of `Thread_ready_mask` set.
In that case it may happen that a de-blocked thread preempts the
currently running thread.

So if the caller knows that no bit of `Thread_ready_mask` is set in the
bits to add to the state, the caller can be sure that this function does
not return `true`.

This change improves coverage.

Change-Id: I29b3d0152c99e7e7127e5e8343af334987113681