modify_senders: fix cursor inconsistency with many threads
If many senders are waiting for a receiver, modify_threads() is preempts
after processing a certain number of receivers. That number is actually
arbitrary chosen and is currently 50.

A cursor is used to remember the next sender required to process. We
have to reset the cursor after all senders were processed because the
cursor remains unset (0) if there were less than 50 senders waiting for
the receiver.

The value of the cursor is used to decide if another modify_senders()
operation may start on this receiver. Without this fix, this operation
can be only performed once if more than 50 senders are waiting for a
receiver but there is no restriction for up to 50 senders.

Fix this inconsistency by always resetting the cursor when it was set
before and the function is left.

Change-Id: I1f5ccfd8132e48655f6e987f68b49424e71b0b28