map_util: Add missing frame unlock when dst is equal to src
In the case that the src and dst of a map operation refer to the same
mapping frame, we did not unlock the frame after unmapping the dst.
The thread that executes the map operation therefore retains ownership
of the frame's lock, even after the map operation has been completed.

If another thread later tries to acquire this frame lock, it becomes
blocked indefinitely, i.e. it tries to help the lock owner thread until
it releases the lock, which of course will never happen (unless, by
chance, the lock owner thread performs another map operation on the
frame), because it does not know that it is holding the lock.

Change-Id: I30929fd0ed3717ea51773167ff20ef7fe32df3bc