arm32: fix non-virt vCPU kernel mode switch
One path to vCPU kernel mode is via leave_by_vcpu_upcall() which is
called indirectly by a continuation. On arm32 the continuation
manipulates the return frame to jump back to svc-mode. This requires
Thread::prepare_vcpu_return_to_kernel() to reset the return frame PSR
to user mode.

This is not an issue on arm64 because the continuation is executed
directly on the kernel exit path instead of manipulating the return
frame.

See ticket #CD-1133.

Change-Id: I7c02d7eb35e638df6f27afb1cac1b8d56b1813ba