dequeue from sender list before transferring IPC items
It may happen that the IPC sender vanishes during the IPC. This happens
when the sender was originally blocked because the receiver was not
ready to receive and later the receiver got ready and started the IPC
transfer on the receiver CPU in the context of the sender.

In Thread::ipc_send_msg() the sender is removed from the receiver`s
sender list _after the transfer of the IPC message has finished. This is
the reason why the  `Abt_ipc_in_progress` path in Thread::abort_send()
is never executed because Sender::in_sender_list() is still true when
this path is reached.

By dequeuing the sender earlier, Thread::abort_send() is already removed
from the sender list when Thread::abort_send() is executed and the case
is properly handled by waiting until the IPC is finished from the
receiver.

Calling recv->vcpu_update_state() handles the case when the sender queue
got empty, see also Receiver::handle_remote_abort_send().

Ipc_sender<Derived>::ipc_send_msg() doesn't transfer message items so
this change is not necessary. But to keep the code consistent, change
the order here as well, including consuming the IRQ.

See ticket #CD-1149.

Change-Id: I08f264df500630391bd6c9a2613e5ed902cf6211