thread: add doorbell interrupt support
When Fiasco should forward hardware interrupts to a guest directly, it
depends on the vCPU state if this is possible. If the thread is
currently in vCPU kernel mode, the VMM must be informed that a guest
interrupt is pending. For this purpose the VMM can register an Irq that
is triggered whenever a new interrupt gets pending.

On Arm GICv4 the hardware provides this feature under the same name. If
the GIC has made a vLPI pending while the vPE is not scheduled, a
doorbell interrupt is raised.

Change-Id: I8acf4e5d39a9f1487cd5633a14858d261601a438