irq: add generic interrupt forwarding code
Add a new "bind_vcpu()" call for Irq_sender objects. The difference is
that, if the interrupt is triggered, it is not received as an IPC
message but rather as guest vIRQ.

This change adds just the generic parts. The exact interrupt
specification is defined by the architecture that supports this
feature. If enabled, the IRQ code will either inject or revoke the vIRQ
from the context as needed.

Change-Id: I91aca9e1d70104ca18760207810febe1822ca08e