arm: add interrupt forwarding support
When an Irq_sender is bound to a vCPU, the kernel will try to allocate
an empty LR in the vGIC to transparently forward the interrupt to the
guest. If no LR is free or if the thread is not in extended vCPU user
mode, the interrupt will stay on a "pending" list. In the latter case
the doorbell interrupt is triggered if configured.

The VMM in user space is expected to not touch any LRs that were not
configured by itself. OTOH the VMM must be prepared to handle EOIs for
vIRQs that were abandoned when detaching active interrupts.

Change-Id: I53e103dd23dc9db0711fbc60908e56726fb4ec13