arm: add direct vtimer vIRQ forwarding
By default the generic timer virtual timer interrupt is passed directly
to the VMM that in turn injects it as proper vIRQ into the guest.
Additionally, Fiasco masks the elapsed vtimer interrupt in the CNTV_CTL
register which actually belongs to the guest. While this works for
Linux, it does not work for various other real time guests.

As an optional optimization, add a vtimer PPI configuration field to
the vCPU state. If enabled, Fiasco will handle the vtimer vIRQ
directly. Because Fiasco then also handles the EOI, there is no need to
fiddle with CNTV_CTL anymore. Instead, Fiasco will mask the vtimer PPI
in the GIC. The PPI will be unmasked as soon as the guest has EOIed the
virtual vtimer interrupt and we re-enter the guest.

Change-Id: I79b12e7a8f1ba9ebe10e0e34f12403022f74f301