riscv: Avoid preprocess function override warning
The preprocess function implementation override functionality is limited
in that it only considers the function name but not the function
arguments. Therefore the function Timer::enable(), which is shared
between architectures and has a default implementation, and the RISC-V
specific Timer::enable(Cpu_number, bool) are considered the same
function by preprocess, which results in a warning.

Change-Id: If3bccbfd17d5d95b66a93d80f594daee7495edef