Optimize IO bitmap flushes via activity tracking
Instead of shooting down the IO bitmaps on all CPUs whenever an IO
port is disabled (which could be used by a malicious thread to
throttle unrelated workloads), we track on which CPUs is the given IO
space active and only flush the CPU IO bitmap on those specific CPUs.

A small drawback of this approach is that the context switching needs
an extra conditional branching in Space::switchin_context() to check
whether the CPU IO bitmap needs to be reset and the resetting needs
an extra atomic operation.

On the other hand, the benefit is that the CPU IO bitmap no longer
needs to be forcefully resynchronized when the IO space stays the same
after the context switch.

Change-Id: Ibe6dbb7cce1f688712b5db4363965e4b467473cd