jdb/arm: gracefully handle not-implemented system registers
The 'kc' debugger command is executed during a debug report. This may
trigger an undefined instruction hyp fault when trying to access a
non-implemented system register.

Handle this condition gracefully: The fault handler enters the nested
trap handler which checks for this condition. The code for dumping the
register content checks if sysreg_fail_pc was reset and prints "n/a" in
that case.

Without this fix, the 'kc' command stops at the first unknown system
register -- and this leaves the debugger report incomplete.

See ticket #CD-1160.

Change-Id: If64b9f4f1c6c300c823459a754be3d973672fdd8