x86: Support arbitrary number of CPUs via TSS/IO bitmap overhaul
Previously, the upper limit of supported CPUs on x86 (approx. 182) was
determined by the fact that the Task State Segments (TSS) for all CPUs
shared a common IO bitmap which had to be mapped within a 64-KiB range
of each of the TSS.

This overhaul introduces a separate IO bitmap for each TSS (i.e. for
each CPU). This enables to support an arbitrary number of CPUs. The fact
that the IO bitmaps are always mapped simplifies the kernel code
substantially.

Implementation details:

* The Generic_io_space class has been reimplemented using the generic
  Bitmap_storage class.

* The IO bitmap of the IO space is synchronized with the CPU IO bitmaps
  on demand (i.e. on an IO port access exception) using a revision
  counter.

* To avoid reseting every bit of the CPU IO bitmap on every context
  switch, the CPU IO bitmap is simply made inaccessible by placing the
  IO bitmap offset beyond the limit of the TSS segment.

  This forces the resynchronization of the CPU IO bitmap with the IO
  bitmap of the IO space via an exception handler on the first IO port
  access. This assumes that the IO accesses are relatively uncommon.

* The superfluous references to IO bitmap page faults on non-x86
  architectures have been removed.

* Since the physical memory for the TSSs is allocated at boot and never
  freed, the allocation is done directly from the memory descriptors.
  This avoids affecting the kernel buddy allocator.

Change-Id: I4a6eb03ce8884357b795c0bd0ce5a82cf699ae70