arm/32: allow to return from certain hyp traps with [debug]
Certain traps are normally not expected to be triggered from hyp mode.
If this happens anyway, hyp_mode_fault() would either halt the kernel
(without debugger) or call the nested trap handler (with debugger).

However, the kernel debugger could trigger an instruction abort when
trying to read from a non-existent system register. In that case the
debugger will remember this condition and skip the instruction so that
the faulting code in the debugger can properly handle this condition.

Change-Id: I44ddecffe88de30d9f4b2b8c625ca8ce30ca0e2d