Space::find_ku_mem: a KU memory region might end at the upper AS limit
If an existing KU memory region ends at the upper AS limit, the next
byte after the region is located at the beginning of the AS.

Consider this by testing against the address of the last byte of the
region instead of testing against the address of the first byte after
the region.

This change fixes an edge case in finding an existing KU memory area.

Change-Id: Ifeb86d8c52758b6d7503196a86679b24a0c130f8