jdb: add handle_early_debug_traps()
The default implementation of this function just returns false. If this
function is overridden by an architecture and returns true, the kernel
debugger is not entered but left right away.

This mechanism can be used to handle traps triggered by the debugger
itself, for example, access to an unimplemented system registers or
similar.

Change-Id: Ia6ba881afa160ecc22e8f493756ad1077d7fef13