arm: fix alien syscalls and exception IPC
The exception IPC wrongly points to the next instruction instead of the
faulting one. This is especially problematic because on arm32 the
instruction length is not fixed. To unify with the page fault path,
always point to the faulting instruction in exception IPC too. This has
implications to multiple kernel paths.

This mainly affects (para-)virtualisation because the trapped SVC/HVC
address is now pointing at the instruction. Interestingly, there was a
corresponding "fixup" in the cpu_virt "undefined instruction" exception
path. Unfortunately the logic there did not handle thumb mode
instructions correctly. Therefore we remove the fixup there too and let
the fault address always point to the trapped instruction. This implies
to adapt the !cpu_virt paths accordingly.

On arm64 and on arm32 with ESR traps (read: cpu_virt), the alien
syscall handling forgot to adjust the PC after the syscall insn was
trapped. This would only trap the pre-syscall state. But as soon as the
exception handler replies, execution resumed after the syscall
instruction. Effectively, this skipped the execution of the syscall
instead of executing it.

Also unify the arm32 syscall entry path if triggered by the "undefined
instruction" exception. It did not support the alien syscall stuff to
begin with.

See ticket #CD-1006.

Change-Id: I6f1695a2c02b3fcf673ff302bcca886401b13f29