VMX: Store the software VMCS fields more compactly
No change of ABI or functionality. This is a preparatory step for
adding a dirty bitmap to the software VMCS which requires a more
compact storage of the software VMCS fields (that allows the bitmap
to fit in).

The physical representation of the offset table is the same.
Previously, the size of each block was the maximum of the sizes of all
blocks. Now each block occupies only as much space as it is actually
needed.

The offset generation code is constexpr now to enable the kernel
setters and getters to use constant values in the generated code in the
future instead of accessing the offset table at run time in the kernel.

The documentation of the structures has also been improved.

Change-Id: I39d9abcfefd56339c4f1d02d9fce38d629e2639f