arm32: fix entry path for undefined instruction
When the kernel is entered through the "undefined instruction" vector,
the return address is only stored in the "PC" field of the return frame.
The supervisor mode LR (SVC_LR) is completely unrelated. Likewise,
overwriting the supervisor mode LR with the adjusted return address is
wrong because this LR belongs to a different mode than the one that
entered the kernel in the first place.

Change-Id: Icbdc2451d0fa896ec057271372f7f396422040ea