spin_lock: Default initialize in unlocked state
Previously Spin_lock had to be initialized explicitly by passing
`Spin_lock<>::Unlocked` to the constructor or by invoking
`Spin_lock::init()`.

Leaving the Spin_lock uninitialized by default is dangerous, as it
is easy to forget the explicit initialization. In fact, there are
some places where Spin_lock is not explicitly initialized, i.e. they
rely on the memory allocated for Spin_lock to be zero initialized.

Therefore, default initialize Spin_lock to be in unlocked state.

While at it, mark copy constructor and copy assignment operator as
deleted to prevent accidental copying of a Spin_lock, as this should
never be an intended operation.

Change-Id: I3cbeda8bcd7708eecedeb44e06c85eae404349ce