VMX: Optimize kernel accesses to the software VMCS
No change of ABI or functionality. This is an optimization of the
kernel accesses to the software VMCS and a preparatory step for nested
virtualization support.

The addressing of the software VMCS fields is done at compile-time in
the kernel to improve performance. The access methods are also
type-safe and provide additional compile-time checks to make sure that
no host-state information is accidentally exposed to user space. This
closes the ticket #CD-811.

The load_guest_state(), store_guest_state() and store_exit_info()
methods have been also moved to a common location where they can be
reused for the nested VM support. Other parts of the code have been
slightly streamlined.

Change-Id: I0e04d9063209bc42f6a611397f95732ef8a78517