ia32/dmar: Replace incorrect usages existence lock
`Dmar::op_un/bind()` acquire the existence lock of the Dmar_space they
work on to ensure that it does not get deleted. However, acquiring an
existence lock is a potential preemption point. So the object might get
deleted before we are able to acquire its existence lock, if we do not
hold a counted reference to it.

The existence lock was not used here to coordinate concurrent
modification (we use atomic operations for that), so it is sufficient
to replace it with a Ref_ptr (counted reference).

See ticket #CD-1166.

Change-Id: Ie4f8d295cf4d238a444d694a334d37cd855b66d4