thread: Prevent dangling pointer in deletion IRQ
When a thread is killed, a deletion IRQ bound to the thread, if any, is
unregistered. However, it remains possible to again bind a deletion IRQ
to the killed thread. Therefore, when the thread is deleted at some
point later, there could again be a deletion IRQ bound to the thread,
leaving a dangling pointer to the deleted thread in the IRQ's
`Irq_base::pin()` property. The same applies to a thread that has never
been started (threads begin in state Thread_dead).

To prevent such a dangling thread pointer, unregister deletion IRQ
again before the delete, at which point no references to the thread
remain, i.e. it is certain that no deletion IRQ can be bound again.

While at it, use access_once to access pointer to the deletion IRQ and
make sure the deletion IRQ's lock is held during unbind.

See ticket #CD-1168.

Change-Id: Ie2cc75ecc2c344a81762fb1055858e4f4bceda60
Requires-Change-Id: I8719d6f4a3635afedc810f20206527f644ec3264