arm: add kernel/usr MPU regions data structures
Out of simplicitly, the classes are still called "Kpdir" and "Pdir" to
keep existing code working as far as possible.

There are a couple of assumptions regarding the MPU region allocation
in the kernel. Specifically, the kernel itself is always mapped in
region 0. Next, the KIP is always mapped in region 1. The reason is
that there is no "user==ro && kernel==rw" setting on arm MPU regions.
Thus, we have to make the KIP writable when entering the kernel and
make it read-only on exit. To keep things simple, the entry/exit code
can hard code the region number for this.

Also the kernel heap needs to have a dedicated region number. The
reason is that on Armv8-R the MPU regions must not overlap. We allocate
ku_mem from the kernel heap and make it accessible to user space. Thus
the kernel heap region must be constrained to the kernel stack on exit.
Otherwise any access from user space to ku_mem would fault because it
hits two regions: kernel heap and the ku_mem mapping. Again, to keep
the entry/exit code lean, assume a dedicated MPU region index.

Change-Id: If518fd3d1ca82abcd1a82b5e1c1de634ccf3c510