mpu: add generic MPU data structures
The MPU configuration is held in a fixed size Mpu_regions object. The
architecture specific code has to supplement the generic classes to
convert the configuration to the particular register interface and also
implement to update code.

The overall goal of the region management is to keep the number of
regions as small as possible. Therefore, regions are merged whenever
possible.

MPU region operations can affect multiple regions e.g., due to a region
split or merge. All operations return a Mpu_regions_update object that
holds the information which regions were affected. The architecture
code can use this to update only affected registers.

Change-Id: Ic832884cd8eed1cbb348239df484f1508a244c12