arm: add early MPU initialization
The MPU is initialized early in the startup by a regular initcall. Only
the kernel image is created as a first region. Other regions like the
KIP or kernel heap are added in their subsystems later. Note that the
MPU might have been enabled by bootstrap before starting Fiasco. In
this case the setup code must make sure that region 0 remains valid all
the time.

Change-Id: Ie43f582af62051ebda53fa7de8570853cba3b5ee