arm: fix exception code on !cpu_virt vCPU svc traps
When trapping an SVC instruction in AArch32 state, the correct HSR.EC
value is 0x11, regardless of whether Fiasco runs in EL1 or EL2. In the
latter case, the exception syndrome is used naturally because it's the
one reported by the hardware in the HSR register. In the former case,
when Fiasco runs on EL1, the vCPU/alien syscall entry path used a
reserved value for no good reason. Fix it to be in line with the other
possible SVC trap paths.

Change-Id: Ic5fd8584befaa9f067dd967c952b61643e5da2c6