arm: add MPU guest state to context
When the kernel runs at EL2 and the thread is in extended vCPU state,
we have to save the EL1 MPU registers. On arm64 this is only necessary
if the VMM has selected the PMSA stage 1 translation regime.

Despite the arm64 stage 1 translation regime we also let the VMM choose
the VTCR_EL2.NSA bit. This bit controls if non-secure stage 2 output
physical address is still treated as secure. As we configure the MPU to
always have secure access permissions, there are no elevated
privileges that can be obtained by this bit.

Change-Id: I482f5bd2d4637da811c1f7e2ccf2834a4c94d369