ia32: apply workaround for Intel IA32e sysret flaw to return-to-kernel
The Intel CPUs generate a #GP in kernel mode if the target RIP is not
canonical while executing sysret. This leads to kernel code running on
an arbitrary user-defined stack pointer. Enforce a canonical RIP before
executing a sysret instruction.

This change adds the missing enforcement to the path for exiting the
kernel to vCPU kernel mode via sysret.

See ticket #CD-1170.

Change-Id: I2ea53e7b1aa40110fb8fc3102f73938c48ee82e6